quiz results are now working

+++ quizzical/src/main/java/bg/softuni/quizzical/repository/QuizRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface QuizRepository extends JpaRepository<Quiz, Long> {
+public interface QuizRepository extends JpaRepository<Quiz, Integer> {
 
     Optional<Quiz> findFirstByCaption(String caption);
 }
